+++ test/Algebra/Graph/Test/NonEmpty/Graph.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE CPP, OverloadedLists, ViewPatterns #-}
 -----------------------------------------------------------------------------
 -- |
--- Module     : Algebra.Graph.Test.Graph.NonEmpty
+-- Module     : Algebra.Graph.Test.NonEmpty.Graph
 -- Copyright  : (c) Andrey Mokhov 2016-2018
 -- License    : MIT (see the file LICENSE)
 -- Maintainer : andrey.mokhov@gmail.com
@@ -9,9 +9,9 @@
 --
 -- Testsuite for "Algebra.Graph.NonEmpty".
 -----------------------------------------------------------------------------
-module Algebra.Graph.Test.Graph.NonEmpty (
+module Algebra.Graph.Test.NonEmpty.Graph (
     -- * Testsuite
-    testGraphNonEmpty
+    testNonEmptyGraph
   ) where
 
 import Prelude ()
@@ -55,9 +55,9 @@ theorems x y = conjoin
     ,             x <= x + y                    // "Overlay order"
     ,         x + y <= x * y                    // "Overlay-connect order" ]
 
-testGraphNonEmpty :: IO ()
-testGraphNonEmpty = do
-    putStrLn "\n============ Graph.NonEmpty ============"
+testNonEmptyGraph :: IO ()
+testNonEmptyGraph = do
+    putStrLn "\n============ NonEmpty.Graph.============"
     test "Axioms of non-empty graphs"   axioms
     test "Theorems of non-empty graphs" theorems
 
@@ -116,14 +116,14 @@ testGraphNonEmpty = do
     test "((x >>= f) >>= g)    == (x >>= (\\y -> (f y) >>= g))" $ mapSize (min 10) $ \(x :: G) (apply -> f) (apply -> g) ->
           ((x >>= f) >>= g)    == (x >>= (\(y :: Int) -> (f y) >>= (g :: Int -> G)))
 
-    putStrLn $ "\n============ Graph.NonEmpty.toNonEmpty ============"
+    putStrLn $ "\n============ NonEmpty.Graph.toNonEmpty ============"
     test "toNonEmpty empty       == Nothing" $
           toNonEmpty (G.empty :: G.Graph Int) == Nothing
 
     test "toNonEmpty (toGraph x) == Just (x :: NonEmpty.Graph a)" $ \x ->
           toNonEmpty (toGraph x) == Just (x :: G)
 
-    putStrLn $ "\n============ Graph.NonEmpty.vertex ============"
+    putStrLn $ "\n============ NonEmpty.Graph.vertex ============"
     test "hasVertex x (vertex x) == True" $ \(x :: Int) ->
           hasVertex x (vertex x) == True
 
@@ -136,7 +136,7 @@ testGraphNonEmpty = do
     test "size        (vertex x) == 1" $ \(x :: Int) ->
           size        (vertex x) == 1
 
-    putStrLn $ "\n============ Graph.NonEmpty.edge ============"
+    putStrLn $ "\n============ NonEmpty.Graph.edge ============"
     test "edge x y               == connect (vertex x) (vertex y)" $ \(x :: Int) y ->
           edge x y               == connect (vertex x) (vertex y)
 
@@ -152,7 +152,7 @@ testGraphNonEmpty = do
     test "vertexCount (edge 1 2) == 2" $
           vertexCount (edge 1 2 :: G) == 2
 
-    putStrLn $ "\n============ Graph.NonEmpty.overlay ============"
+    putStrLn $ "\n============ NonEmpty.Graph.overlay ============"
     test "hasVertex z (overlay x y) == hasVertex z x || hasVertex z y" $ \(x :: G) y z ->
           hasVertex z (overlay x y) == hasVertex z x || hasVertex z y
 
@@ -177,7 +177,7 @@ testGraphNonEmpty = do
     test "edgeCount   (overlay 1 2) == 0" $
           edgeCount   (overlay 1 2 :: G) == 0
 
-    putStrLn $ "\n============ Graph.NonEmpty.overlay1 ============"
+    putStrLn $ "\n============ NonEmpty.Graph.overlay1 ============"
     test "               overlay1 empty x == x" $ \(x :: G) ->
                          overlay1 G.empty x == x
 
@@ -185,7 +185,7 @@ testGraphNonEmpty = do
           x /= G.empty ==> overlay1 x   y == overlay (fromJust $ toNonEmpty x) y
 
 
-    putStrLn $ "\n============ Graph.NonEmpty.connect ============"
+    putStrLn $ "\n============ NonEmpty.Graph.connect ============"
     test "hasVertex z (connect x y) == hasVertex z x || hasVertex z y" $ \(x :: G) y z ->
           hasVertex z (connect x y) == hasVertex z x || hasVertex z y
 
@@ -216,7 +216,7 @@ testGraphNonEmpty = do
     test "edgeCount   (connect 1 2) == 1" $
           edgeCount   (connect 1 2 :: G) == 1
 
-    putStrLn $ "\n============ Graph.NonEmpty.vertices1 ============"
+    putStrLn $ "\n============ NonEmpty.Graph.vertices1 ============"
     test "vertices1 [x]           == vertex x" $ \(x :: Int) ->
           vertices1 [x]           == vertex x
 
@@ -232,7 +232,7 @@ testGraphNonEmpty = do
         let xs = NonEmpty.fromList (getNonEmpty xs')
         in (vertexSet   . vertices1) xs == (Set.fromList . NonEmpty.toList) xs
 
-    putStrLn $ "\n============ Graph.NonEmpty.edges1 ============"
+    putStrLn $ "\n============ NonEmpty.Graph.edges1 ============"
     test "edges1 [(x,y)]     == edge x y" $ \(x :: Int) y ->
           edges1 [(x,y)]     == edge x y
 
@@ -240,21 +240,21 @@ testGraphNonEmpty = do
         let xs = NonEmpty.fromList (getNonEmpty xs')
         in (edgeCount . edges1) xs == (NonEmpty.length . NonEmpty.nub) xs
 
-    putStrLn $ "\n============ Graph.NonEmpty.overlays1 ============"
+    putStrLn $ "\n============ NonEmpty.Graph.overlays1 ============"
     test "overlays1 [x]   == x" $ \(x :: G) ->
           overlays1 [x]   == x
 
     test "overlays1 [x,y] == overlay x y" $ \(x :: G) y ->
           overlays1 [x,y] == overlay x y
 
-    putStrLn $ "\n============ Graph.NonEmpty.connects1 ============"
+    putStrLn $ "\n============ NonEmpty.Graph.connects1 ============"
     test "connects1 [x]   == x" $ \(x :: G) ->
           connects1 [x]   == x
 
     test "connects1 [x,y] == connect x y" $ \(x :: G) y ->
           connects1 [x,y] == connect x y
 
-    putStrLn $ "\n============ Graph.NonEmpty.foldg1 ============"
+    putStrLn $ "\n============ NonEmpty.Graph.foldg1 ============"
     test "foldg1 vertex    overlay connect        == id" $ \(x :: G) ->
           foldg1 vertex    overlay connect x      == id x
 
@@ -267,7 +267,7 @@ testGraphNonEmpty = do
     test "foldg1 (== x)    (||)    (||)           == hasVertex x" $ \(x :: Int) y ->
           foldg1 (== x)    (||)    (||) y         == hasVertex x y
 
-    putStrLn $ "\n============ Graph.NonEmpty.isSubgraphOf ============"
+    putStrLn $ "\n============ NonEmpty.Graph.isSubgraphOf ============"
     test "isSubgraphOf x             (overlay x y) ==  True" $ \(x :: G) y ->
           isSubgraphOf x             (overlay x y) ==  True
 
@@ -282,7 +282,7 @@ testGraphNonEmpty = do
         let y = x + z -- Make sure we hit the precondition
         in isSubgraphOf x y                        ==> x <= y
 
-    putStrLn "\n============ Graph.NonEmpty.(===) ============"
+    putStrLn "\n============ NonEmpty.Graph.(===) ============"
     test "    x === x     == True" $ \(x :: G) ->
              (x === x)    == True
 
@@ -295,7 +295,7 @@ testGraphNonEmpty = do
     test "x + y === x * y == False" $ \(x :: G) y ->
          (x + y === x * y) == False
 
-    putStrLn $ "\n============ Graph.NonEmpty.size ============"
+    putStrLn $ "\n============ NonEmpty.Graph.size ============"
     test "size (vertex x)    == 1" $ \(x :: Int) ->
           size (vertex x)    == 1
 
@@ -311,14 +311,14 @@ testGraphNonEmpty = do
     test "size x             >= vertexCount x" $ \(x :: G) ->
           size x             >= vertexCount x
 
-    putStrLn $ "\n============ Graph.NonEmpty.hasVertex ============"
+    putStrLn $ "\n============ NonEmpty.Graph.hasVertex ============"
     test "hasVertex x (vertex x) == True" $ \(x :: Int) ->
           hasVertex x (vertex x) == True
 
     test "hasVertex 1 (vertex 2) == False" $
           hasVertex 1 (vertex 2 :: G) == False
 
-    putStrLn $ "\n============ Graph.NonEmpty.hasEdge ============"
+    putStrLn $ "\n============ NonEmpty.Graph.hasEdge ============"
     test "hasEdge x y (vertex z)       == False" $ \(x :: Int) y z ->
           hasEdge x y (vertex z)       == False
 
@@ -332,7 +332,7 @@ testGraphNonEmpty = do
         (u, v) <- elements ((x, y) : edgeList z)
         return $ hasEdge u v z == elem (u, v) (edgeList z)
 
-    putStrLn $ "\n============ Graph.NonEmpty.vertexCount ============"
+    putStrLn $ "\n============ NonEmpty.Graph.vertexCount ============"
     test "vertexCount (vertex x) == 1" $ \(x :: Int) ->
           vertexCount (vertex x) == 1
 
@@ -342,7 +342,7 @@ testGraphNonEmpty = do
     test "vertexCount            == length . vertexList1" $ \(x :: G) ->
           vertexCount x          == (NonEmpty.length . vertexList1) x
 
-    putStrLn $ "\n============ Graph.NonEmpty.edgeCount ============"
+    putStrLn $ "\n============ NonEmpty.Graph.edgeCount ============"
     test "edgeCount (vertex x) == 0" $ \(x :: Int) ->
           edgeCount (vertex x) == 0
 
@@ -352,7 +352,7 @@ testGraphNonEmpty = do
     test "edgeCount            == length . edgeList" $ \(x :: G) ->
           edgeCount x          == (length . edgeList) x
 
-    putStrLn $ "\n============ Graph.NonEmpty.vertexList1 ============"
+    putStrLn $ "\n============ NonEmpty.Graph.vertexList1 ============"
     test "vertexList1 (vertex x)  == [x]" $ \(x :: Int) ->
           vertexList1 (vertex x)  == [x]
 
@@ -360,7 +360,7 @@ testGraphNonEmpty = do
         let xs = NonEmpty.fromList (getNonEmpty xs')
         in (vertexList1 . vertices1) xs == (NonEmpty.nub . NonEmpty.sort) xs
 
-    putStrLn $ "\n============ Graph.NonEmpty.edgeList ============"
+    putStrLn $ "\n============ NonEmpty.Graph.edgeList ============"
     test "edgeList (vertex x)     == []" $ \(x :: Int) ->
           edgeList (vertex x)     == []
 
@@ -377,7 +377,7 @@ testGraphNonEmpty = do
     test "edgeList . transpose    == sort . map swap . edgeList" $ \(x :: G) ->
          (edgeList . transpose) x == (sort . map swap . edgeList) x
 
-    putStrLn $ "\n============ Graph.NonEmpty.vertexSet ============"
+    putStrLn $ "\n============ NonEmpty.Graph.vertexSet ============"
     test "vertexSet . vertex    == Set.singleton" $ \(x :: Int) ->
          (vertexSet . vertex) x == Set.singleton x
 
@@ -389,7 +389,7 @@ testGraphNonEmpty = do
         let xs = NonEmpty.fromList (getNonEmpty xs')
         in (vertexSet . clique1) xs == (Set.fromList . NonEmpty.toList) xs
 
-    putStrLn $ "\n============ Graph.NonEmpty.edgeSet ============"
+    putStrLn $ "\n============ NonEmpty.Graph.edgeSet ============"
     test "edgeSet (vertex x) == Set.empty" $ \(x :: Int) ->
           edgeSet (vertex x) == Set.empty
 
@@ -400,7 +400,7 @@ testGraphNonEmpty = do
         let xs = NonEmpty.fromList (getNonEmpty xs')
         in (edgeSet . edges1) xs == (Set.fromList . NonEmpty.toList) xs
 
-    putStrLn $ "\n============ Graph.NonEmpty.path1 ============"
+    putStrLn $ "\n============ NonEmpty.Graph.path1 ============"
     test "path1 [x]       == vertex x" $ \(x :: Int) ->
           path1 [x]       == vertex x
 
@@ -411,7 +411,7 @@ testGraphNonEmpty = do
         let xs = NonEmpty.fromList (getNonEmpty xs')
         in (path1 . NonEmpty.reverse) xs == (transpose . path1) xs
 
-    putStrLn $ "\n============ Graph.NonEmpty.circuit1 ============"
+    putStrLn $ "\n============ NonEmpty.Graph.circuit1 ============"
     test "circuit1 [x]       == edge x x" $ \(x :: Int) ->
           circuit1 [x]       == edge x x
 
@@ -422,7 +422,7 @@ testGraphNonEmpty = do
         let xs = NonEmpty.fromList (getNonEmpty xs')
         in (circuit1 . NonEmpty.reverse) xs == (transpose . circuit1) xs
 
-    putStrLn $ "\n============ Graph.NonEmpty.clique1 ============"
+    putStrLn $ "\n============ NonEmpty.Graph.clique1 ============"
     test "clique1 [x]        == vertex x" $ \(x :: Int) ->
           clique1 [x]        == vertex x
 
@@ -441,7 +441,7 @@ testGraphNonEmpty = do
         let xs = NonEmpty.fromList (getNonEmpty xs')
         in (clique1 . NonEmpty.reverse) xs == (transpose . clique1) xs
 
-    putStrLn $ "\n============ Graph.NonEmpty.biclique1 ============"
+    putStrLn $ "\n============ NonEmpty.Graph.biclique1 ============"
     test "biclique1 [x1,x2] [y1,y2] == edges1 [(x1,y1), (x1,y2), (x2,y1), (x2,y2)]" $ \(x1 :: Int) x2 y1 y2 ->
           biclique1 [x1,x2] [y1,y2] == edges1 [(x1,y1), (x1,y2), (x2,y1), (x2,y2)]
 
@@ -450,7 +450,7 @@ testGraphNonEmpty = do
             ys = NonEmpty.fromList (getNonEmpty ys')
         in biclique1 xs      ys      == connect (vertices1 xs) (vertices1 ys)
 
-    putStrLn $ "\n============ Graph.NonEmpty.star ============"
+    putStrLn $ "\n============ NonEmpty.Graph.star ============"
     test "star x []    == vertex x" $ \(x :: Int) ->
           star x []    == vertex x
 
@@ -460,7 +460,7 @@ testGraphNonEmpty = do
     test "star x [y,z] == edges1 [(x,y), (x,z)]" $ \(x :: Int) y z ->
           star x [y,z] == edges1 [(x,y), (x,z)]
 
-    putStrLn $ "\n============ Graph.NonEmpty.stars1 ============"
+    putStrLn $ "\n============ NonEmpty.Graph.stars1 ============"
     test "stars1 [(x, [] )]               == vertex x" $ \(x :: Int) ->
           stars1 [(x, [] )]               == vertex x
 
@@ -479,7 +479,7 @@ testGraphNonEmpty = do
           ys = NonEmpty.fromList (getNonEmpty ys')
       in  overlay (stars1 xs) (stars1 ys) == stars1 (xs <> ys)
 
-    putStrLn $ "\n============ Graph.NonEmpty.tree ============"
+    putStrLn $ "\n============ NonEmpty.Graph.tree ============"
     test "tree (Node x [])                                         == vertex x" $ \(x :: Int) ->
           tree (Node x [])                                         == vertex x
 
@@ -492,7 +492,7 @@ testGraphNonEmpty = do
     test "tree (Node 1 [Node 2 [], Node 3 [Node 4 [], Node 5 []]]) == edges1 [(1,2), (1,3), (3,4), (3,5)]" $
           tree (Node 1 [Node 2 [], Node 3 [Node 4 [], Node 5 []]]) == edges1 [(1,2), (1,3), (3,4), (3,5::Int)]
 
-    putStrLn $ "\n============ Graph.NonEmpty.mesh1 ============"
+    putStrLn $ "\n============ NonEmpty.Graph.mesh1 ============"
     test "mesh1 [x]     [y]        == vertex (x, y)" $ \(x :: Int) (y :: Int) ->
           mesh1 [x]     [y]        == vertex (x, y)
 
@@ -512,7 +512,7 @@ testGraphNonEmpty = do
             ys = NonEmpty.fromList (getNonEmpty ys')
          in size (mesh1 xs ys) == max 1 (3 * length xs * length ys - length xs - length ys -1)
 
-    putStrLn $ "\n============ Graph.NonEmpty.torus1 ============"
+    putStrLn $ "\n============ NonEmpty.Graph.torus1 ============"
     test "torus1 [x]   [y]        == edge (x,y) (x,y)" $ \(x :: Int) (y :: Int) ->
           torus1 [x]   [y]        == edge (x,y) (x,y)
 
@@ -532,7 +532,7 @@ testGraphNonEmpty = do
             ys = NonEmpty.fromList (getNonEmpty ys')
         in size (torus1 xs ys) == max 1 (3 * length xs * length ys)
 
-    putStrLn $ "\n============ Graph.NonEmpty.removeVertex1 ============"
+    putStrLn $ "\n============ NonEmpty.Graph.removeVertex1 ============"
     test "removeVertex1 x (vertex x)          == Nothing" $ \(x :: Int) ->
           removeVertex1 x (vertex x)          == Nothing
 
@@ -548,7 +548,7 @@ testGraphNonEmpty = do
     test "removeVertex1 x >=> removeVertex1 x == removeVertex1 x" $ \(x :: Int) y ->
          (removeVertex1 x >=> removeVertex1 x) y == removeVertex1 x y
 
-    putStrLn $ "\n============ Graph.NonEmpty.removeEdge ============"
+    putStrLn $ "\n============ NonEmpty.Graph.removeEdge ============"
     test "removeEdge x y (edge x y)       == vertices1 [x,y]" $ \(x :: Int) y ->
           removeEdge x y (edge x y)       == vertices1 [x,y]
 
@@ -564,7 +564,7 @@ testGraphNonEmpty = do
     test "size (removeEdge x y z)         <= 3 * size z" $ \(x :: Int) y z ->
           size (removeEdge x y z)         <= 3 * size z
 
-    putStrLn $ "\n============ Graph.NonEmpty.replaceVertex ============"
+    putStrLn $ "\n============ NonEmpty.Graph.replaceVertex ============"
     test "replaceVertex x x            == id" $ \(x :: Int) y ->
           replaceVertex x x y          == y
 
@@ -574,7 +574,7 @@ testGraphNonEmpty = do
     test "replaceVertex x y            == mergeVertices (== x) y" $ \(x :: Int) y z ->
           replaceVertex x y z          == mergeVertices (== x) y z
 
-    putStrLn $ "\n============ Graph.NonEmpty.mergeVertices ============"
+    putStrLn $ "\n============ NonEmpty.Graph.mergeVertices ============"
     test "mergeVertices (const False) x    == id" $ \(x :: Int) y ->
           mergeVertices (const False) x y  == y
 
@@ -587,7 +587,7 @@ testGraphNonEmpty = do
     test "mergeVertices odd  1 (3 + 4 * 5) == 4 * 1" $
           mergeVertices odd  1 (3 + 4 * 5) == (4 * 1 :: G)
 
-    putStrLn $ "\n============ Graph.NonEmpty.splitVertex1 ============"
+    putStrLn $ "\n============ NonEmpty.Graph.splitVertex1 ============"
     test "splitVertex1 x [x]                 == id" $ \x (y :: G) ->
           splitVertex1 x [x] y               == y
 
@@ -597,7 +597,7 @@ testGraphNonEmpty = do
     test "splitVertex1 1 [0,1] $ 1 * (2 + 3) == (0 + 1) * (2 + 3)" $
           splitVertex1 1 [0,1] (1 * (2 + 3)) == (0 + 1) * (2 + 3 :: G)
 
-    putStrLn $ "\n============ Graph.NonEmpty.transpose ============"
+    putStrLn $ "\n============ NonEmpty.Graph.transpose ============"
     test "transpose (vertex x)  == vertex x" $ \(x :: Int) ->
           transpose (vertex x)  == vertex x
 
@@ -613,7 +613,7 @@ testGraphNonEmpty = do
     test "edgeList . transpose  == sort . map swap . edgeList" $ \(x :: G) ->
          (edgeList . transpose) x == (sort . map swap . edgeList) x
 
-    putStrLn $ "\n============ Graph.NonEmpty.induce1 ============"
+    putStrLn $ "\n============ NonEmpty.Graph.induce1 ============"
     test "induce1 (const True ) x == Just x" $ \(x :: G) ->
           induce1 (const True ) x == Just x
 
@@ -626,7 +626,7 @@ testGraphNonEmpty = do
     test "induce1 p >=> induce1 q == induce1 (\\x -> p x && q x)" $ \(apply -> p) (apply -> q) (y :: G) ->
          (induce1 p >=> induce1 q) y == induce1 (\x -> p x && q x) y
 
-    putStrLn $ "\n============ Graph.NonEmpty.simplify ============"
+    putStrLn $ "\n============ NonEmpty.Graph.simplify ============"
     test "simplify             ==  id" $ \(x :: G) ->
           simplify x           ==  x
 
@@ -645,7 +645,7 @@ testGraphNonEmpty = do
     test "simplify (1 * 1 * 1) === 1 * 1" $
           simplify (1 * 1 * 1) === (1 * 1 :: G)
 
-    putStrLn "\n============ Graph.NonEmpty.sparsify ============"
+    putStrLn "\n============ NonEmpty.Graph.sparsify ============"
     test "sort . reachable x       == sort . rights . reachable (Right x) . sparsify" $ \x (y :: G) ->
          (sort . reachable x) y    == (sort . rights . reachable (Right x) . sparsify) y
 
@@ -658,7 +658,7 @@ testGraphNonEmpty = do
     test "size        (sparsify x) <= 3 * size x" $ \(x :: G) ->
           size        (sparsify x) <= 3 * size x
 
-    putStrLn "\n============ Graph.NonEmpty.box ============"
+    putStrLn "\n============ NonEmpty.Graph.box ============"
     test "box (path1 [0,1]) (path1 ['a','b']) == <correct result>" $ mapSize (min 10) $
           box (path1 [0,1]) (path1 ['a','b']) == edges1 [ ((0,'a'), (0,'b'))
                                                         , ((0,'a'), (1,'a'))