+++ test/Main.hs
@@ -1,10 +1,10 @@
 import Algebra.Graph.Test.AdjacencyIntMap
 import Algebra.Graph.Test.AdjacencyMap
-import Algebra.Graph.Test.AdjacencyMap.NonEmpty
+import Algebra.Graph.Test.NonEmpty.AdjacencyMap
 import Algebra.Graph.Test.Export
 import Algebra.Graph.Test.Fold
 import Algebra.Graph.Test.Graph
-import Algebra.Graph.Test.Graph.NonEmpty
+import Algebra.Graph.Test.NonEmpty.Graph
 import Algebra.Graph.Test.Internal
 import Algebra.Graph.Test.Labelled.AdjacencyMap
 import Algebra.Graph.Test.Relation
@@ -14,12 +14,12 @@ main :: IO ()
 main = do
     testAdjacencyIntMap
     testAdjacencyMap
-    testAdjacencyMapNonEmpty
     testExport
     testFold
     testGraph
-    testGraphNonEmpty
     testInternal
     testLabelledAdjacencyMap
+    testNonEmptyAdjacencyMap
+    testNonEmptyGraph
     testRelation
     testTyped