+++ CHANGES.md
@@ -3,7 +3,7 @@
 ## 0.3
 
 * #136: Rename `Algebra.Graph.NonEmpty.NonEmptyGraph` to `Algebra.Graph.NonEmpty.Graph`.
-* #136: Add `Algebra.Graph.AdjacencyMap.NonEmpty`.
+* #136: Add `Algebra.Graph.NonEmpty.AdjacencyMap`.
 * #136: Remove `vertexIntSet` from the API of basic graph data types. Also
         remove `Algebra.Graph.adjacencyMap` and `Algebra.Graph.adjacencyIntMap`.
         This functionality is still available from the type class `ToGraph`.